refactor(crypto): split noise internals into separate file
[no changelog]
