+++ core/SConscript.prodtest
@@ -109,7 +109,8 @@ SOURCE_MOD_CRYPTO += [
     'vendor/trezor-crypto/hmac.c',
     'vendor/trezor-crypto/hmac_drbg.c',
     'vendor/trezor-crypto/memzero.c',
-    'vendor/trezor-crypto/noise.c',
+    'vendor/trezor-crypto/noise_common.c',
+    'vendor/trezor-crypto/noise_kk1.c',
     'vendor/trezor-crypto/nist256p1.c',
     'vendor/trezor-crypto/rand.c',
     'vendor/trezor-crypto/ripemd160.c',
