+++ core/SConscript.prodtest_emu
@@ -91,7 +91,8 @@ SOURCE_MOD_CRYPTO += [
     'vendor/trezor-crypto/hmac.c',
     'vendor/trezor-crypto/hmac_drbg.c',
     'vendor/trezor-crypto/memzero.c',
-    'vendor/trezor-crypto/noise.c',
+    'vendor/trezor-crypto/noise_common.c',
+    'vendor/trezor-crypto/noise_kk1.c',
     'vendor/trezor-crypto/nist256p1.c',
     'vendor/trezor-crypto/rand.c',
     'vendor/trezor-crypto/rand_insecure.c',
