+++ core/embed/projects/prodtest/cmd/secure_channel.h
@@ -19,10 +19,10 @@
 
 #pragma once
 
-#include "noise.h"
+#include "noise_kk1.h"
 
-#define SECURE_CHANNEL_INPUT_SIZE (sizeof(noise_response_t))
-#define SECURE_CHANNEL_OUTPUT_SIZE (sizeof(noise_request_t))
+#define SECURE_CHANNEL_INPUT_SIZE (sizeof(noise_kk1_response_t))
+#define SECURE_CHANNEL_OUTPUT_SIZE (sizeof(noise_kk1_request_t))
 #define SECURE_CHANNEL_TAG_SIZE (NOISE_TAG_SIZE)
 
 bool secure_channel_handshake_1(uint8_t output[SECURE_CHANNEL_OUTPUT_SIZE]);
